Swap in the new permissions system as the main authentication system
