+++ h/security/permits.py
@@ -1,42 +1,35 @@
-import warnings
-from logging import getLogger
-from typing import Optional
-
-from pyramid.authorization import ACLAuthorizationPolicy
-
-from h.security import permission_map
 from h.security.identity import Identity
-from h.security.principals import principals_for_identity
-
-LOG = getLogger(__name__)
+from h.security.permission_map import PERMISSION_MAP
 
 
-def identity_permits(identity: Optional[Identity], context, permission) -> bool:
+def identity_permits(identity: Identity, context, permission) -> bool:
     """
-    Get whether a given identity has the requested permission on the context.
+    Check whether a given identity has permission to operate on a context.
 
-    :param identity: Identity object or None for non authenticated access
-    :param context: A context object
-    :param permission: The permission requested
+    For example the identity might include a user, and the context a group
+    and the permission might ask whether the user can edit that group.
+
+    :param identity: Identity object of the user
+    :param context: Context object representing the objects acted upon
+    :param permission: Permission requested
     """
-
-    acl_allows = bool(
-        ACLAuthorizationPolicy().permits(
-            context=context,
-            principals=principals_for_identity(identity),
-            permission=permission,
+    if clauses := PERMISSION_MAP.get(permission):
+        # Grant the permissions if for *any* single clause...
+        return any(
+            # .. *all* elements in it are true
+            all(_predicate_true(predicate, identity, context) for predicate in clause)
+            for clause in clauses
         )
-    )
 
+    return False
+
+
+def _predicate_true(predicate, identity, context):
+    """Check whether a predicate is true."""
     try:
-        map_allows = permission_map.identity_permits(identity, context, permission)
-    # pylint: disable=broad-except
-    except Exception as err:  # pragma: no cover
-        map_allows = err  # pylint: disable=redefined-variable-type
+        return predicate(identity, context)
 
-    if map_allows != acl_allows:  # pragma: no cover
-        message = f"Permissions system disagree about {permission}: ACL={acl_allows}, MAP={map_allows}"
-        warnings.warn(message, PendingDeprecationWarning)
-        LOG.warning(message)
-
-    return acl_allows
+    # If the "predicate" isn't callable, we'll assume it's a permission
+    # and work out if we have that permission
+    except TypeError:
+        return identity_permits(identity, context, predicate)
